Moved styles into the separate folder
